fix(soma): corrige a funcionalidade de soma
Agora as contas podem ser feitas com decimais, não apenas inteiros

+++ src/soma.h
@@ -1,6 +1,6 @@
 #ifndef SOMA_H
 #define SOMA_H
 
-int soma(int num1, int num2);
+float soma(float num1, float num2);
 
 #endif
